+++ packages/cinterop/src/jvm/kotlin/io/realm/internal/interop/sync/SyncErrorCodeCategory.kt
@@ -3,11 +3,11 @@ package io.realm.internal.interop.sync
 import io.realm.internal.interop.realm_sync_error_category_e
 
 actual enum class SyncErrorCodeCategory(val nativeValue: Int) {
-    CLIENT(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_CLIENT),
-    CONNECTION(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_CONNECTION),
-    SESSION(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_SESSION),
-    SYSTEM(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_SYSTEM),
-    UNKNOWN(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_UNKNOWN);
+    RLM_SYNC_ERROR_CATEGORY_CLIENT(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_CLIENT),
+    RLM_SYNC_ERROR_CATEGORY_CONNECTION(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_CONNECTION),
+    RLM_SYNC_ERROR_CATEGORY_SESSION(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_SESSION),
+    RLM_SYNC_ERROR_CATEGORY_SYSTEM(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_SYSTEM),
+    RLM_SYNC_ERROR_CATEGORY_UNKNOWN(realm_sync_error_category_e.RLM_SYNC_ERROR_CATEGORY_UNKNOWN);
 
     companion object {
         @JvmStatic